requirements.txt
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
drf-yasg==1.17.0
gunicorn==19.9.0
psycopg2==2.7.5
requests==2.23.0
